+++ app/controllers/dorsale/billing_machine/invoices_controller.rb
@@ -39,6 +39,7 @@ class Dorsale::BillingMachine::InvoicesController < ::Dorsale::BillingMachine::A
     authorize model, :create?
 
     if @invoice.save
+      Dorsale::BillingMachine::PdfFileGenerator.(@invoice)
       flash[:notice] = t("messages.invoices.create_ok")
       redirect_to default_back_url
     else
@@ -78,6 +79,7 @@ class Dorsale::BillingMachine::InvoicesController < ::Dorsale::BillingMachine::A
     authorize @invoice, :update?
 
     if @invoice.update(invoice_params_for_update)
+      Dorsale::BillingMachine::PdfFileGenerator.(@invoice)
       flash[:notice] = t("messages.invoices.update_ok")
       redirect_to default_back_url
     else