+++ app/uploaders/dorsale/pdf_uploader.rb
@@ -1,0 +1,5 @@
+class Dorsale::PdfUploader < ::Dorsale::ApplicationUploader
+  def extension_white_list
+    %w(pdf)
+  end
+end