+++ lib/dorsale/alexandrie/prawn.rb
@@ -4,7 +4,7 @@ module Dorsale
       def render_with_attachments
         final_pdf = ::CombinePDF.parse(self.render)
 
-        @main_document.attachments.each do |attachment|
+        attachments.each do |attachment|
           next unless File.extname(attachment.file.path) == ".pdf"
 
           final_pdf << ::CombinePDF.load(attachment.file.path)