+++ spec/pdfs/dorsale/billing_machine/invoice_single_vat_pdf_spec.rb
@@ -16,12 +16,14 @@ describe ::Dorsale::BillingMachine::InvoiceSingleVatPdf, pdfs: true do
     i
   }
 
+  let(:generate!) {
+    Dorsale::BillingMachine::PdfFileGenerator.(invoice)
+    invoice.reload
+  }
+
   let(:content) {
-    tempfile = Tempfile.new("pdf")
-    tempfile.binmode
-    tempfile.write(invoice.to_pdf)
-    tempfile.flush
-    Yomu.new(tempfile.path).text
+    generate!
+    Yomu.new(invoice.pdf_file.path).text
   }
 
   it "should display global vat rate" do
@@ -30,12 +32,15 @@ describe ::Dorsale::BillingMachine::InvoiceSingleVatPdf, pdfs: true do
     expect(content).to_not include "TVA %"
   end
 
-  it "should work with empty invoice" do
-    id_card = Dorsale::BillingMachine::IdCard.new
-    invoice = ::Dorsale::BillingMachine::Invoice.new(id_card: id_card)
+  describe "empty invoice" do
+    let(:invoice) {
+      id_card = Dorsale::BillingMachine::IdCard.new
+      id_card.save(validate: false)
+      invoice = ::Dorsale::BillingMachine::Invoice.create!(id_card: id_card)
+    }
 
-    expect {
-      invoice.to_pdf
-    }.to_not raise_error
-  end
+    it "should work" do
+      expect { generate! }.to_not raise_error
+    end
+  end # describe "empty invoice"
 end