+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Delete Flyboy folders
 - Add `link_to_object` helper
 - Change/fix expense states (new -> draft, submited -> submitted)
+- Persist invoice/quotation PDFs after save
 
 ## 3.3.0